Stage 2
* Revert using sets for router collections
* Introduce Node utility module to hold utility functions, move some
  there
* Find outbound stream in dialback by id
